requirements.txt
Flask==2.3.3
python3-saml==1.15.0
ldap3==2.9.1
requests==2.31.0
cryptography==41.0.4
xmlsec
lxml>=4.9.0
python-dotenv==1.0.0
pyjwt==2.8.0
bcrypt==4.0.1
gunicorn==21.2.0